Add Lift

+++ src/States/EndResult.cpp
@@ -81,7 +81,7 @@ void EndResult::drawStats()
 
     std::string time = "TIME:";
     DrawTextEx(font, time.c_str(), {511, 571}, 40, 2, WHITE);   
-    int totalSec = StatsManager::getInstance().time_taken; // small epsilon for safety
+    int totalSec = (int)StatsManager::getInstance().time_taken; // small epsilon for safety
     int mins = totalSec / 60;
     int secs = totalSec % 60;
